칵테일 평점, 칵테일 한글 이름 null 예외 처리

+++ backend/src/main/java/com/ssafy/cocktail/backend/cocktails/dto/CocktailLikesInterface.java
@@ -5,6 +5,6 @@ public interface CocktailLikesInterface {
     String getCocktailNameKo();
     String getCocktailNameEn();
     String getCocktailImg();
-    float getCocktailRating();
+    Double getCocktailRating();
     Long getCnt();
 }
